Add patch to replace compiler version with template
This will allow the lockdir to expect a compiler of a given version. The
goal is to allow whatever version of ocaml happens to be installed on
the system to build Alice to avoid needing to build the compiler as a
package when building Alice in source-based package managers such as
homebrew.